Refactor (APIMF-3491): extracted key value property parser
